+++ packages/uni_ui/lib/cards/widgets/restaurant_menu_item.dart
@@ -1,0 +1,35 @@
+import 'package:flutter/material.dart';
+import 'package:phosphor_flutter/phosphor_flutter.dart';
+
+class RestaurantMenuItem extends StatelessWidget {
+  const RestaurantMenuItem({
+    super.key,
+    required this.name,
+    required this.icon});
+
+  final String name;
+  final PhosphorIcon icon;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      padding: const EdgeInsets.symmetric(horizontal: 0, vertical: 8.0),
+      child: Row(
+        crossAxisAlignment: CrossAxisAlignment.center,
+        children: [
+          Expanded(
+              flex: 1,
+              child: icon),
+          Expanded(
+            flex: 5,
+            child: Text(
+              name,
+              style: Theme.of(context).textTheme.bodyLarge,
+              softWrap: true,
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
